requirements.txt
appdirs==1.4.4
arrow==1.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
beautifulsoup4==4.9.3
blinker==1.4
brotlipy==0.7.0
bs4==0.0.1
cached-property==1.5.2
certifi==2020.6.20
cloudpickle==1.6.0
coloredlogs==15.0.1
cssselect==1.1.0
cycler==0.10.0
docopt==0.6.2
fake-useragent==0.1.11
feedparser==6.0.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
gym==0.18.3
h5py==2.10.0
humanfriendly==9.2
ipython-genutils==0.2.0
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
mkl-fft==1.2.0
mkl-random==1.0.4
mkl-service==2.3.0
mpl-finance==0.10.1
mplfinance==0.12.7a17
msgpack==1.0.2
msgpack-numpy==0.4.7.1
multitasking==0.0.9
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
opencensus-context==0.1.2
packaging==21.0
pandas==1.2.4
parse==1.19.0
parso==0.8.2
patsy==0.5.1
Pillow==8.2.0
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.1.0
pyglet==1.5.15
PyJWT==1.7.1
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pyreadline==2.1
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests-html==0.10.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
seaborn==0.11.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
statsmodels==0.12.2
stocktrends==0.1.5
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
Tensorforce==0.6.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
tqdm==4.61.0
traitlets==5.0.5
urllib3==1.26.5
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
xgboost==1.4.2
yahoo-fin==0.8.8
yfinance==0.1.63

